requirements.txt
django
pydantic-settings
python-telegram-bot
psycopg2-binary
requests